requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==24.4.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.0
deepl==1.18.0
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.0.3
django-autoslug==1.9.9
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-decouple==2.1
django-filter==24.2
django-modeltranslation==0.19.4
django-redis==5.4.0
django-tinymce==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
fcm-django==0.3.11
fernet==1.0.1
firebase-admin==6.5.0
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
msgpack==1.0.8
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyfcm==1.4.9
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyotp==2.9.0
pyparsing==3.1.2
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.3
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
service-identity==24.1.0
setuptools==70.1.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.6.0
zope.interface==6.4.post2
